Move results functions out of client.py
