+++ clients/python/tests/test_get_results.py
@@ -4,7 +4,7 @@ import json
 from gql import Client
 
 # Avoid confusion with gql Client
-import tracker_client.client as tc
+import tracker_client.results as tc
 import tracker_client.queries as queries
 
 # pylint: disable=no-member
@@ -13,7 +13,7 @@ import tracker_client.queries as queries
 def test_get_all_results_error(mocker, error_message):
     """Test that get_all_results correctly handles error response"""
     mocker.patch(
-        "tracker_client.client.execute_query",
+        "tracker_client.results.execute_query",
         auto_spec=True,
         return_value=error_message,
     )
@@ -30,7 +30,7 @@ def test_get_all_results_error(mocker, error_message):
 def test_get_all_results(mocker, all_results_input, all_results_output):
     """Test that get_all_results produces correct output"""
     mocker.patch(
-        "tracker_client.client.execute_query",
+        "tracker_client.results.execute_query",
         auto_spec=True,
         return_value=all_results_input,
     )
@@ -48,7 +48,7 @@ def test_get_all_results(mocker, all_results_input, all_results_output):
 def test_get_web_results_error(mocker, error_message):
     """Test that get_web_results correctly handles error response"""
     mocker.patch(
-        "tracker_client.client.execute_query",
+        "tracker_client.results.execute_query",
         auto_spec=True,
         return_value=error_message,
     )
@@ -65,7 +65,7 @@ def test_get_web_results_error(mocker, error_message):
 def test_get_web_results(mocker, web_results_input, web_results_output):
     """Test that get_web_results produces correct output"""
     mocker.patch(
-        "tracker_client.client.execute_query",
+        "tracker_client.results.execute_query",
         auto_spec=True,
         return_value=web_results_input,
     )
@@ -83,7 +83,7 @@ def test_get_web_results(mocker, web_results_input, web_results_output):
 def test_get_email_results_error(mocker, error_message):
     """Test that get_email_results correctly handles error response"""
     mocker.patch(
-        "tracker_client.client.execute_query",
+        "tracker_client.results.execute_query",
         auto_spec=True,
         return_value=error_message,
     )
@@ -100,7 +100,7 @@ def test_get_email_results_error(mocker, error_message):
 def test_get_email_results(mocker, email_results_input, email_results_output):
     """Test that get_email_results produces correct output"""
     mocker.patch(
-        "tracker_client.client.execute_query",
+        "tracker_client.results.execute_query",
         auto_spec=True,
         return_value=email_results_input,
     )
@@ -118,7 +118,7 @@ def test_get_email_results(mocker, email_results_input, email_results_output):
 def test_get_domain_status_error(mocker, error_message):
     """Test that get_domain_status correctly handles error response"""
     mocker.patch(
-        "tracker_client.client.execute_query",
+        "tracker_client.results.execute_query",
         auto_spec=True,
         return_value=error_message,
     )
@@ -135,7 +135,7 @@ def test_get_domain_status_error(mocker, error_message):
 def test_get_domain_status(mocker, domain_status_input, domain_status_output):
     """Test that get_domain_status produces correct output"""
     mocker.patch(
-        "tracker_client.client.execute_query",
+        "tracker_client.results.execute_query",
         auto_spec=True,
         return_value=domain_status_input,
     )
